Update bonus.cpp
Changed salary input type to float.

+++ Krati Agrawal/If-Else/bonus.cpp
@@ -2,14 +2,14 @@
 using namespace std;
 int main()
 {
-    int salary, years;
+    float salary, years;
     cout << "Enter Salary : ";
     cin >> salary;
     cout << "Enter Years of Service : ";
     cin >> years;
     if (years > 5)
     {
-        int bonus = (salary * 5) / 100;
+        float bonus = (salary * 5) / 100;
         cout << "The net bonus amount is" << bonus << endl;
     }
     else
